// apps/aitoearn-channel/src/libs/twitter/twitter.interfaces.ts
export interface XMediaUploadInitRequest {
  // command: 'INIT' | 'APPEND' | 'FINALIZE'
  media_type: XMediaType
  total_bytes: number
  media_category: XMediaCategory
  shared: boolean
}
// ...
export interface XChunkedMediaUploadRequest {
  media: Blob
  media_id: string
  segment_index: number
}

export interface Geo {
  place_id: string
}

export interface PostMedia {
  media_ids: string[]
  tagged_users?: string[]
}

enum PostPollReplySettings {
  // following, mentionedUsers, subscribers, verified
  FOLLOWING = 'following',
  MENTIONED_USERS = 'mentionedUsers',
  SUBSCRIBERS = 'subscribers',
  VERIFIED = 'verified',
}

export interface PostPoll {
  options: string[]
  duration_minutes: number
  reply_settings?: PostPollReplySettings
}

export interface postReply {
  in_reply_to_tweet_id: string
  exclude_reply_user_ids?: string[]
}
export interface XCreatePostRequest {
  card_uri?: string
  community_id?: string
  direct_message_deep_link?: string
  for_super_followers_only?: boolean
  geo?: Geo
  media?: PostMedia
  nullcast?: boolean
  poll?: PostPoll
  quote_tweet_id?: string
  reply?: postReply
  reply_settings?: PostPollReplySettings
  text?: string
}